requirements.txt
fastapi
fastapi-cache2
uvicorn
redis
openapi-core
aioitertools
pydantic==1.10.10